feature/VIH-5331 : Work In Progress
