+++ AdminWebsite/AdminWebsite/ClientApp/src/app/get-audio-file/get-audio-link-button/get-audio-link-button.component.ts
@@ -1,5 +1,7 @@
-import { Component } from '@angular/core';
-import { AudioLinkService, AudioLinkState } from '../../services/audio-link-service';
+import { Component, Input } from '@angular/core';
+import { AudioLinkService } from '../../services/audio-link-service';
+import { AudioLinkState } from '../../services/audio-link-state';
+import { Logger } from '../../services/logger';
 
 @Component({
     selector: 'app-get-audio-link-button',
@@ -7,15 +9,24 @@ import { AudioLinkService, AudioLinkState } from '../../services/audio-link-serv
     styleUrls: ['./get-audio-link-button.component.scss']
 })
 export class GetAudioLinkButtonComponent {
-    buttonText: string;
+    public audioLinkStates: typeof AudioLinkState = AudioLinkState;
+    private _currentLinkRetrievalState: AudioLinkState = AudioLinkState.finished;
 
-    constructor(private audioLinkService: AudioLinkService) {
-        this.buttonText = audioLinkService.currentLinkRetrievalState.toString();
-    }
+    @Input() hearingId: string;
+
+    constructor(private audioLinkService: AudioLinkService, private logger: Logger) {}
 
     async onGetLinkClick() {
-        const audioLink = await this.audioLinkService.getAudioLink('sdfs');
-        this.buttonText = audioLink;
+        try {
+            this.setCurrentState(AudioLinkState.loading);
+
+            const audioLink = await this.audioLinkService.getAudioLink('sdfs');
+
+            this.setCurrentState(AudioLinkState.finished);
+        } catch (error) {
+            this.logger.error(`Error retrieving audio link for: ${this.hearingId}`, error);
+            this.setCurrentState(AudioLinkState.error);
+        }
     }
 
     async onCopyLinkClick() {
@@ -23,6 +34,14 @@ export class GetAudioLinkButtonComponent {
     }
 
     showOnState(audioLinkState: AudioLinkState) {
-        return this.audioLinkService.currentLinkRetrievalState;
+        return audioLinkState === this._currentLinkRetrievalState;
+    }
+
+    get currentLinkRetrievalState() {
+        return this._currentLinkRetrievalState;
+    }
+
+    setCurrentState(audioLinkState: AudioLinkState) {
+        this._currentLinkRetrievalState = audioLinkState;
     }
 }
